Add framework project and begin porting sections.

+++ src/Shared/Caching/ICacheHash.cs
@@ -14,6 +14,6 @@ namespace SixLabors.ImageSharp.Web.Caching
         /// <param name="value">The input value to hash</param>
         /// <param name="length">The length of the returned hash without any extensions</param>
         /// <returns>The <see cref="string"/></returns>
-        string Create(string value, uint length);
+        string Create(string value, byte length);
     }
 }
